inventory-api/inventory-api/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=23.5,<24.0
drf-spectacular>=0.27,<0.28
typing-extensions>=4.12,<5.0
